requirements.txt
# Data and Dataset Management
kagglehub
datasetsfastapi
uvicorn
python-multipart
Pillow
